+++ lib/auto_html.rb
@@ -1,10 +1,17 @@
 %w(base filter builder auto_html_for).each do |f|
-  require File.dirname(__FILE__) + "/auto_html/#{f}"
+  require File.expand_path("../auto_html/#{f}", __FILE__)
 end
 
-# Register built-in filters
 Dir["#{File.dirname(__FILE__) + '/auto_html/filters'}/**/*"].each do |filter|
   require "#{filter}"
 end
 
-ActiveRecord::Base.send :include, AutoHtmlFor
+# if rails
+require 'auto_html/railtie' if defined?(Rails::Railtie)
+if defined?(ActiveRecord::Base)
+  ActiveRecord::Base.send :include, AutoHtmlFor
+
+  module ActionView::Helpers::TextHelper
+    include AutoHtml
+  end
+end