+++ lib/auto_html/base.rb
@@ -1,4 +1,6 @@
 module AutoHtml
+  extend self
+
   def self.add_filter(name, &block)
     AutoHtml::Builder.add_filter(name, &block)
   end
@@ -6,6 +8,11 @@ module AutoHtml
   def auto_html(raw, &proc)
     return "" if raw.blank?
     builder = Builder.new(raw)
-    builder.instance_eval(&proc)
+    result = builder.instance_eval(&proc)
+    return raw if result.nil?
+    result.respond_to?(:html_safe) ?
+      result.html_safe :
+        result
   end
+
 end