+++ lib/auto_html/filters/image.rb
@@ -1,5 +1,15 @@
-AutoHtml.add_filter(:image) do |text|
-  text.gsub(/http:\/\/.+\.(jpg|jpeg|bmp|gif|png)(\?\S+)?/i) do |match|
-    %|<img src="#{match}" alt=""/>|
+require 'redcarpet'
+
+class NoParagraphRenderer < ::Redcarpet::Render::XHTML
+  def paragraph(text)
+    text
+  end    
+end
+
+AutoHtml.add_filter(:image).with({:alt => ''}) do |text, options|
+  r = Redcarpet::Markdown.new(NoParagraphRenderer)
+  alt = options[:alt]
+  text.gsub(/https?:\/\/.+?\.(jpg|jpeg|bmp|gif|png)(\?\S+)?/i) do |match|
+    r.render("![#{alt}](#{match})")
   end
 end