+++ lib/auto_html/filters/link.rb
@@ -1,5 +1,7 @@
-require 'action_view'
+require 'rinku'
+require 'rexml/document'
 
 AutoHtml.add_filter(:link).with({}) do |text, options|
-  ActionView::Base.new.auto_link(text, :html => options, :sanitize => false)
+  attributes = Array(options).reject { |k,v| v.nil? }.map { |k, v| %{#{k}="#{REXML::Text::normalize(v)}"} }.join(' ')
+  Rinku.auto_link(text, :all, attributes)
 end