+++ lib/auto_html/filters/youtube.rb
@@ -1,12 +1,14 @@
-AutoHtml.add_filter(:youtube).with(:width => 390, :height => 250, :frameborder => 0) do |text, options|
-  regex = /http:\/\/(www.)?youtube\.com\/watch\?v=([A-Za-z0-9._%-]*)(\&\S+)?/
+AutoHtml.add_filter(:youtube).with(:width => 420, :height => 315, :frameborder => 0, :wmode => nil) do |text, options|
+  regex = /https?:\/\/(www.)?(youtube\.com\/watch\?v=|youtu\.be\/|youtube\.com\/watch\?feature=player_embedded&v=)([A-Za-z0-9_-]*)(\&\S+)?(\S)*/
   text.gsub(regex) do
-    youtube_id = $2
+    youtube_id = $3
     width = options[:width]
     height = options[:height]
     frameborder = options[:frameborder]
-    %{<iframe class="youtube-player" type="text/html" width="#{width}" height="#{height}" src="http://www.youtube.com/embed/#{youtube_id}" frameborder="#{frameborder}">
-</iframe>}
+		wmode = options[:wmode]
+		src = "//www.youtube.com/embed/#{youtube_id}"
+		src += "?wmode=#{wmode}" if wmode
+    %{<iframe width="#{width}" height="#{height}" src="#{src}" frameborder="#{frameborder}" allowfullscreen></iframe>}
   end
 end