+++ test/functional/auto_html_for_options_test.rb
@@ -1,24 +1,24 @@
-require File.dirname(__FILE__) + '/../test_helper'
-require File.dirname(__FILE__) + '/../fixture_setup'
+require File.expand_path('../../test_helper', __FILE__)
+require File.expand_path('../../fixture_setup', __FILE__)
 
 # store default so we can revert so that other tests use default option
 default_suffix = AutoHtmlFor.auto_html_for_options[:htmlized_attribute_suffix]
 AutoHtmlFor.auto_html_for_options[:htmlized_attribute_suffix] = '_htmlized'
 
-class Article < ActiveRecord::Base
-  auto_html_for :body do
+class Post < ActiveRecord::Base
+  auto_html_for :content do
     simple_format
   end
 end
 
-class AutoHtmlForTest < Test::Unit::TestCase
+class AutoHtmlForOptionsTest < Test::Unit::TestCase
   include FixtureSetup
 
   def test_transform_after_save
-    @article = Article.new(:body => 'Yo!')
-    assert_equal '<p>Yo!</p>', @article.body_htmlized
+    @article = Post.new(:content => 'Yo!')
+    assert_equal '<p>Yo!</p>', @article.content_htmlized
     @article.save!
-    assert_equal '<p>Yo!</p>', @article.body_htmlized
+    assert_equal '<p>Yo!</p>', @article.content_htmlized
   end
 end